+++ src/main/java/kro/dodoworld/advancedmonsters/command/tab/AbilityTabCompleter.java
@@ -1,7 +1,7 @@
 package kro.dodoworld.advancedmonsters.command.tab;
 
 import kro.dodoworld.advancedmonsters.util.AdvancedUtils;
-import kro.dodoworld.advancedmonsters.util.MonsterAbility;
+import kro.dodoworld.advancedmonsters.modifier.ability.MonsterAbility;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabCompleter;
