+++ src/main/java/kro/dodoworld/advancedmonsters/modifier/EntityModifier.java
@@ -9,9 +9,9 @@ import kro.dodoworld.advancedmonsters.config.modifier.StormyModifierConfig;
 import kro.dodoworld.advancedmonsters.config.modifier.TankModifierConfig;
 import kro.dodoworld.advancedmonsters.modifier.ability.type.RevitalizeModifier;
 import kro.dodoworld.advancedmonsters.modifier.ability.type.TeleporterModifier;
-import kro.dodoworld.advancedmonsters.modifier.ai.AnimalAttackTargetGoal;
+import kro.dodoworld.advancedmonsters.modifier.entity.ai.AnimalAttackTargetGoal;
 import kro.dodoworld.advancedmonsters.util.AdvancedUtils;
-import kro.dodoworld.advancedmonsters.util.MonsterAbility;
+import kro.dodoworld.advancedmonsters.modifier.ability.MonsterAbility;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextDecoration;
