+++ src/main/java/kro/dodoworld/advancedmonsters/modifier/entity/equipment/weapon/EntityBow.java
@@ -1,0 +1,4 @@
+package kro.dodoworld.advancedmonsters.modifier.entity.equipment.weapon;
+
+public class EntityBow {
+}
