--- src/main/java/kro/dodoworld/advancedmonsters/modifier/equipment/weapon/EntityBow.java
@@ -1,4 +1,0 @@
-package kro.dodoworld.advancedmonsters.modifier.equipment.weapon;
-
-public class EntityBow {
-}
